refactoring some shared logic into AbstractDataType class... more work to do
