+++ src/DataModel/DataType/NullType.php
@@ -1,0 +1,80 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\DataModel\DataType;
+
+use App\DataModel\Contracts\AbstractDataType;
+use App\DataModel\Contracts\DataTypeInterface;
+use App\DataModel\Contracts\SerializableInterface;
+use App\DataModel\Translation\LanguageCodes;
+
+class NullType extends AbstractDataType
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function setValue($val, array $options = []): DataTypeInterface
+    {
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getValue(array $options = [])
+    {
+        return null;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function isNull(array $options = []): bool
+    {
+        return true;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getLanguages(): array
+    {
+        return [LanguageCodes::ANY];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getSerializationId(): string
+    {
+        return 'NULL';
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function toArray(): array
+    {
+        return [
+            'type' => $this->getSerializationId(),
+            'val' => null,
+        ];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function fromArray(array $data): SerializableInterface
+    {
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function canMergeWith(DataTypeInterface $type): bool
+    {
+        return $type instanceof NullType;
+    }
+}
