+++ src/DataModel/DataType/NumericType.php
@@ -2,30 +2,23 @@
 
 declare(strict_types=1);
 
-namespace App\DataModel\Types;
+namespace App\DataModel\DataType;
 
+use App\DataModel\Contracts\AbstractDataType;
+use App\DataModel\Contracts\DataTypeInterface;
 use App\DataModel\Contracts\SerializableInterface;
-use App\DataModel\Contracts\TypeInterface;
-use App\DataModel\Contracts\VersionableTrait;
 use App\DataModel\Translation\LanguageCodes;
 use App\Exception\ErrorMessages;
 use App\Exception\WanderlusterException;
 
-class MimeType implements TypeInterface
+class NumericType extends AbstractDataType
 {
-    use VersionableTrait;
-
-    const PATTERN = '/^[-\w]+\/[-\w\.\+]+$/';
+    protected ?float $val = null;
 
     /**
-     * @var string|null
-     */
-    protected $val;
-
-    /**
-     * MimeType constructor.
+     * Numeric constructor.
      *
-     * @param string|null $val
+     * @param float|int|null $val
      *
      * @throws WanderlusterException
      */
@@ -42,7 +35,7 @@ class MimeType implements TypeInterface
      */
     public function getSerializationId(): string
     {
-        return 'MIME_TYPE';
+        return 'NUM';
     }
 
     /**
@@ -85,16 +78,10 @@ class MimeType implements TypeInterface
     /**
      * {@inheritdoc}
      */
-    public function setValue($val, array $options = []): TypeInterface
+    public function setValue($val, array $options = []): DataTypeInterface
     {
-        if (!is_string($val) && !is_null($val)) {
-            throw new WanderlusterException(sprintf(ErrorMessages::INVALID_DATATYPE_VALUE, $this->getSerializationId(), 'String required'));
-        }
-
-        if (is_string($val)) {
-            if (!preg_match(self::PATTERN, $val)) {
-                throw new WanderlusterException(sprintf(ErrorMessages::INVALID_DATATYPE_VALUE, $this->getSerializationId(), 'Invalid MimeType'));
-            }
+        if (!is_int($val) && !is_float($val) && !is_null($val)) {
+            throw new WanderlusterException(sprintf(ErrorMessages::INVALID_DATATYPE_VALUE, $this->getSerializationId(), 'Numeric required'));
         }
 
         $this->val = $val;
@@ -126,35 +113,11 @@ class MimeType implements TypeInterface
         return [LanguageCodes::ANY];
     }
 
-    public function merge(TypeInterface $type): void
+    /**
+     * {@inheritdoc}
+     */
+    public function canMergeWith(DataTypeInterface $type): bool
     {
-        if (!$type instanceof MimeType) {
-            throw new WanderlusterException(sprintf(ErrorMessages::MERGE_UNSUCCESSFUL, $type->getSerializationId(), $this->getSerializationId()));
-        }
-
-        $thisVal = $this->getValue();
-        $thatVal = $type->getValue();
-        $thisVer = $this->getVersion();
-        $thatVer = $type->getVersion();
-
-        // previous version... do nothing
-        if ($thatVer < $thatVer) {
-            return;
-        }
-
-        // greater version, use its value
-        if ($thatVer > $thisVer) {
-            $this->setVersion($thatVer);
-            $this->setValue($thatVal);
-
-            return;
-        }
-
-        // handle merge conflict
-        if ($thatVer === $thisVer && $thisVal !== $thatVal) {
-            if ($thatVal > $thisVal) {
-                $this->setValue($thatVal);
-            }
-        }
+        return $type instanceof NumericType;
     }
 }
