+++ tests/DataModel/DataType/DateTimeTypeTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\DataModel\Types;
+namespace App\Tests\DataModel\DataType;
 
-use App\DataModel\Types\BooleanType;
-use App\DataModel\Types\DateTimeType;
+use App\DataModel\DataType\BooleanType;
+use App\DataModel\DataType\DateTimeType;
 use App\Exception\WanderlusterException;
 use DateTime;
 use DateTimeImmutable;
