+++ tests/DataModel/DataType/FileSizeTypeTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\DataModel\Types;
+namespace App\Tests\DataModel\DataType;
 
-use App\DataModel\Types\BooleanType;
-use App\DataModel\Types\FileSizeType;
+use App\DataModel\DataType\BooleanType;
+use App\DataModel\DataType\FileSizeType;
 use App\Exception\WanderlusterException;
 use PHPUnit\Framework\TestCase;
